requirements.txt
django
djangorestframework
django-debug-toolbar
pyjwt
psycopg2-binary
black
flake8
django-cors-headers